Moved testing file.
Because the testing file was inside the `/src` folder, it was included
in build processes.
